backend/requirements.txt
# 🧠 MOTOR DE PREDIÇÕES REAL - Dependências atualizadas

# Core FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de dados
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.8  # Comentado por problemas de compilação
# asyncpg==0.29.0  # Comentado por problemas de compilação

# Cache e Redis
redis==5.0.1
aioredis==2.0.1

# Configuração e validação
pydantic==2.11.9
pydantic-settings==2.10.1
python-dotenv==1.1.1

# HTTP client para APIs
httpx==0.25.2
requests==2.31.0

# 📊 ANÁLISE MATEMÁTICA E ESTATÍSTICA
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2

# 🧮 Machine Learning - Sistema Avançado v2.0
joblib==1.3.2
# tensorflow==2.15.0  # Comentado por ser pesado - descomentar se necessário

# Processamento e parsing
beautifulsoup4==4.12.2
python-multipart==0.0.6
aiofiles==23.2.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSockets e real-time
python-socketio==5.10.0
websockets==12.0

# Background tasks
celery==5.3.4
apscheduler==3.10.4
schedule==1.2.0

# 🧪 Testes
pytest==7.4.3
pytest-asyncio==0.21.1

# 📡 Dependências específicas para APIs de futebol
# (Todas as dependências já estão listadas acima)